// backend/src/types/request/dedicatedPage/insertDedicatedPageWayToEarnRequest.ts
export type InsertDedicatedPageWayToEarnRequest = {
  general_background_color: string;
  general_title_text: string;
  general_subtitle_text: string;
  general_title_font_size: number;
  general_title_font_color: string;
  general_subtitle_font_size: number;
  general_subtitle_font_color: string;
  status: string;
  user_id: number;
  admin_ref: number;
  created_by: number;
};