// backend/src/types/request/dedicatedPage/updateDedicatedPageWayToEarnRequest.ts
export type UpdateDedicatedPageWayToEarnRequest = {
  general_background_color: string;
  general_title_text: string;
  general_subtitle_text: string;
  general_title_font_size: number;
  general_title_font_color: string;
  general_subtitle_font_size: number;
  general_subtitle_font_color: string;
  status: string;
  admin_ref: number;
  updated_by: number;
};